Refactor composer model to be separate from base transformer
